Add ability to configure OAuth Service in Multi User Che
