+++ assembly/assembly-wsmaster-war/src/main/webapp/WEB-INF/classes/che/multiuser.properties
@@ -136,3 +136,6 @@ che.keycloak.js_adapter_url=NULL
 #     a discovery endpoint as detailed in the following specification
 #     https://openid.net/specs/openid-connect-discovery-1_0.html#ProviderConfig
 che.keycloak.oidc_provider=NULL
+
+#
+che.oauth.type=che
