+++ multiuser/keycloak/che-multiuser-keycloak-server/src/main/java/org/eclipse/che/multiuser/keycloak/server/deploy/KeycloakModule.java
@@ -19,7 +19,7 @@ import org.eclipse.che.multiuser.keycloak.server.KeycloakConfigurationService;
 import org.eclipse.che.multiuser.keycloak.server.KeycloakTokenValidator;
 import org.eclipse.che.multiuser.keycloak.server.KeycloakUserManager;
 import org.eclipse.che.multiuser.keycloak.server.dao.KeycloakProfileDao;
-import org.eclipse.che.multiuser.keycloak.server.oauth2.KeycloakOAuthAuthenticationService;
+import org.eclipse.che.security.oauth.OAuthAuthenticationService;
 
 public class KeycloakModule extends AbstractModule {
   @Override
@@ -29,9 +29,12 @@ public class KeycloakModule extends AbstractModule {
         .to(org.eclipse.che.multiuser.keycloak.server.KeycloakHttpJsonRequestFactory.class);
     bind(TokenValidator.class).to(KeycloakTokenValidator.class);
     bind(KeycloakConfigurationService.class);
-    bind(KeycloakOAuthAuthenticationService.class);
 
     bind(ProfileDao.class).to(KeycloakProfileDao.class);
     bind(PersonalAccountUserManager.class).to(KeycloakUserManager.class);
+
+    bind(OAuthAuthenticationService.class)
+        .toProvider(OAuthAuthenticationServiceProvider.class)
+        .asEagerSingleton();
   }
 }
